frontend/requirements.txt
flask==3.0.0
google-cloud-run==0.10.0
google-cloud-storage==2.14.0
gunicorn==21.2.0
